removed trainings reducer and moved logic to calendar also deleted unused and bad code preparing to refactoring training summary and logic with crud trainings
