+++ src/pages/diary/Diary.js
@@ -1,25 +1,20 @@
-import React, {useEffect, useState} from 'react';
+import React, {useEffect} from 'react';
 import moment from 'moment';
 import {connect} from 'react-redux';
-import {initCalendar, trainingToShowHandler, deleteTraining} from "../../store/actions/trainings.actions";
+import {initCalendar, deleteTraining} from "../../store/actions/calendar-trainings.actions";
 import TrainingSummary from "./components/TrainingSummary";
 import Calendar from "./components/Calendar";
 import Spinner from "../../components/ui/spinner/Spinner";
-import Backdrop from "../../components/ui/back-drop/Backdrop";
 import './Diary.scss';
 import MonthPicker from "./components/MonthPicker";
 
 const Diary = (props) => {
   const {
     initCalendar,
-    trainingToShow,
-    trainingToShowHandler,
     error,
-    deleteTraining,
-    trainingToDelete,
-    calendarStructureLength
+    calendarStructureLength,
+    loading,
   } = props;
-  const [modal, setModal] = useState(false);
   const today = moment().format('YYYY-MM-DD');
 
   useEffect(() => {
@@ -27,16 +22,7 @@ const Diary = (props) => {
     initCalendar(today);
   }, [initCalendar, today, calendarStructureLength]);
 
-  if (error === 404) {
-    return (
-      <h2 className="diary__warning">
-        Brak treningów w danym miesiacu! <br/>
-        Dodaj swój pierwszy trening w tym miesiacu!
-      </h2>
-    );
-  }
-
-  if (error) {
+  if (error && error !== 404) {
     return (
       <h2 style={{textAlign: 'center', marginTop: '70px', fontSize: '24px'}}>
         Coś poszło nie tak!
@@ -44,58 +30,22 @@ const Diary = (props) => {
     );
   }
 
-  const handleDeleteTraining = () => {
-    if (trainingToShow.length === 2) {
-      deleteTraining(trainingToDelete[0].id);
-      setModal(false)
-    } else {
-      deleteTraining(trainingToShow[0].id);
-      setModal(false)
-    }
-  };
-
-  let diary = <Spinner/>;
-  let popUp = (
-    <div className="popUp" style={{
-      transform: modal ? 'translateY(0)' : 'translateY(-100vh)',
-      opacity: modal ? '1' : '0'
-    }}>
-      <h2>Na pewno chcesz usunąć trening?</h2>
-      <button type="button" onClick={() => setModal(false)} style={{color: 'green'}}>
-        Anuluj
-      </button>
-
-      <button type="button" onClick={handleDeleteTraining} style={{color: '#bd2130'}}>
-        Usuń trening
-      </button>
-    </div>
-  );
-
-  if (trainingToShow) {
-    diary = (
+  return (
+    loading ? <Spinner/> :
       <>
-        <Backdrop show={modal} cancel={() => setModal(false)}/>
-        {popUp}
         <MonthPicker/>
         <Calendar/>
-        <TrainingSummary setModal={setModal}/>
       </>
-    );
-  }
-
-  return diary;
+  );
 };
 
 const mapStateToProps = state => {
   return {
     token: state.auth.token,
-    trainings: state.trainings.trainings,
-    trainingToShow: state.trainings.trainingToShow,
-    error: state.trainings.error,
-    loading: state.trainings.loading,
-    trainingToDelete: state.trainings.trainingToDelete,
+    error: state.calendar.error,
+    loading: state.calendar.loading,
     calendarStructureLength: state.calendar.calendarStructure.length,
   }
 };
 
-export default connect(mapStateToProps, {initCalendar, trainingToShowHandler, deleteTraining})(Diary);
+export default connect(mapStateToProps, {initCalendar, deleteTraining})(Diary);
