+++ src/store/action-types.js
@@ -1,10 +1,7 @@
 export const LOADING = 'LOADING';
 export const GET_TRAININGS_SUCCESS = 'GET_TRAININGS_SUCCESS';
 export const GET_TRAININGS_FAIL = 'GET_TRAININGS_FAIL';
-export const TRAINING_TO_SHOW_HANDLER = 'TRAINING_TO_SHOW_HANDLER';
-export const TRAININGS_CLEAR_ERROR = 'TRAININGS_CLEAR_ERROR';
-export const EDIT_TRAINING_IN_STORE = 'EDIT_TRAINING_IN_STORE';
-export const TRAINING_TO_DELETE = 'TRAINING_TO_DELETE';
+export const EDIT_TRAINING_IN_CALENDAR = 'EDIT_TRAINING_IN_CALENDAR';
 
 export const ADD_TRAINING_STEP_1 = 'ADD_TRAINING_STEP_1';
 export const ADD_TRAINING_EDIT_FORM = 'ADD_TRAINING_EDIT_FORM';
